Add date sorting and add column display text
Added 'date' as a possible sorting method.  This is , in theory, even
slower than naturalized, but it should support date formats in both
European and American integer date formats (Like "1/20/2015" or
"20-1-2015").  I may write a full parser for this, but right now all I
feel the need to support is the above.  CSS changes will be next on my
list.